fix: Munki prefs shouldn't have ruby nils
When ruby parses the attributes for preferences it doesn't strip the
`nil` values and instead will write the value as <data> inside the
profile.

Example:
    <key>RecoveryKeyFile</key>
    <data>
BAgw
    </data>

This will render as `<040830>` inside of System Preferences.

This causes Munki to use this malformed data and is unable to
interpret it properly causing https://github.com/munki/munki/issues/799.
Munki can be more resilient to bad data but Chef shouldn't create the bad
data in the first place.