Use raw Firefox AutoConfig instead of CCK2.
CCK2 broke for a while on Firefox because of the changes introduced with
Quantum. This cookbook isn't doing anything complex with CCK2, so
convert it to use raw Firefox AutoConfig instead to insulate against the
next time Firefox makes a breaking change to their API and extensions.

Fixes facebook/IT-CPE#171