Adding new helpers library
Summary:

As part of our Tech debt week and ongoing effort to both clean up Chef code and make more of it open-source, we need to port a bunch of the helper functions from cpe_utils over to cpe_helpers. These are things that should be "public" - i.e. things that query the system or system state and are not unique to FB infra.

In this case, if we're going to make cpe_helpers a first-class citizen, we need to ensure that it has the functions we need from cpe_utils.

Since we so far only have node functions in cpe_helpers, I created a new library called cpe_helpers, `CPE::Helpers`, that will contain non-node functions that can be used in other library functions, and I've ported over a number of previously-only-in-cpe_utils functions to it.

The major change here is that the console_user and related functions are now library functions and not part of node, which makes it easy to use them in other libraries. As a result, there's been a bit of refactoring to remove the use of node.

Specifically:
* loginctl_users no longer makes a distinction of Debian vs. Fedora, instead choosing to go with Fedora as the default value. We can revisit this later if we come up with a non-node-based version of making this distinction.
* console_user no longer considers /etc/fb-machine-owner when checking for the current logged in user. We'll need to come up with another method if the existing ones are not working for Fedora.

Reviewed By: mikedodge04

Differential Revision: D17403520

fbshipit-source-id: 4878e375afd1b1ee91f8b3119cd593dcc886a50a