Don't manage /etc/paths inside cpe_pathsd (#220)
Summary:

`/etc/paths` should be managed outside of this API (imo).

This API is arbitrarily managing the system `paths` file for macOS but not Linux.

As system paths can change, it should be left up to the org to decide if they want to manage this or not, especially since it has no impact on cpe_pathsd itself.

In fact, Apple did change system paths in catalina:
```
/usr/local/bin
/usr/bin
/bin
/usr/sbin
/sbin
/Library/Apple/usr/bin
/Library/Apple/bin
```

The alternative would be to add to the API to allow path injection into the system `paths` file (at the top).
Pull Request resolved: https://github.com/facebook/IT-CPE/pull/220

Reviewed By: nmcspadden

Differential Revision: D18102345

Pulled By: achand

fbshipit-source-id: 4874543851758bac599b29e1b3627202085fa6ae