minimally handle github issue with cpe_chrome windows
Summary:
We don't actually use this setting internally for anything so we've never seen if it actually works. The short of it is there are some Chrome settings that need to be `:string` types in the registry that are actually just JSON strings.

The resource can't really do that without explicitly setting the setting to be a JSON string, for example:
```
node.default['cpe_chrome']['profile']['ManagedBookmarks'] = [
  {
    'name' => 'My Bookmark',
    'url' => 'https://google.com'
  },
  {
    'name' => 'Another Bookmark',
    'url' => 'https://google.com'
  }
].to_json
```

To maintain the API we have I will add support to render the appropriate `ManagedBookmarks` setting to the registry. Additionally, a unit test rendering the example setting in the [corresponding documentation](https://cloud.google.com/docs/chrome-enterprise/policies/?policy=ManagedBookmarks) is added to ensure the output is valid.

Reviewed By: chilcote

Differential Revision: D18846631

fbshipit-source-id: dc88d96a5150b61f574ae4bd28aae824e44c696b