RFC: turn on update alerts for CPE Macs
Summary:

Add two new keys to the Chrome prefs to help keep Chrome up to date by utilizing the built-in prompts:

https://support.google.com/chrome/a/answer/7679871?hl=en
86400000 is 24 hours

While I'm waiting for my test to work out (I still haven't seen this supposed prompt), thought I'd put this up for discussion. The idea here is to help ensure folks are on latest Chrome, which will help reduce the potential for SEVs when there's a vulnerability (i.e. when MSC forced updates for macOS as well as Chrome).

Reviewed By: DNSGeek

Differential Revision: D18979758

fbshipit-source-id: c826c6be89ec3e0049b88e11a999ba1f9572ca04