ignore certain prefs for canary
Summary:

Want to not apply the relaunch requirement for chrome canary. This diff creates a new node attribute, `canary_ignored_prefs`, and when the profile is created it will ignore keys in this list.

Also, adding these pref keys to that attributes file, so we need to also add them to the windows registry library.

Reviewed By: DNSGeek

Differential Revision: D19339139

fbshipit-source-id: 3e1ce752ef899193cde22f621c66234e878f6f0e