simplify loginctl parsing
Summary:

Ideally we can use `loginctl -o json list-users` to get the user listing in a machine-readable form; however, this was broken until very recent versions of `systemd`. Instead, use `loginctl --no-legend` which already remove the cruft we don't care about.

Change the output of `CPE::Helpers.loginctl_users` to match that of `loginctl -o json list-users` so we can switch to it at some point.

```
~
❯ loginctl -o json list-users
[{"uid":1001,"user":"michel"}]
```

Reviewed By: svmastersamurai

Differential Revision: D19433035

fbshipit-source-id: 4eb25db961b04c3aaa2ca3fee24deeaa19b857f8