also install kernel
Summary:

The UX for `cpe_kernel_channel` is better if we can have it automate the kernel installation, but in a way that does not impose too much overhead. We can't rely on checking if the specific package N-V-R (name-version-release) is installed because we might want to say "any 4.18.0 kernel from CentOS".

Since Fedora and CentOS kernels install the kernel image in a known location, check that to see if the specified kernel is installed by globbing the expected location, and install it otherwise.

Reviewed By: steelcowboy

Differential Revision: D20073286

fbshipit-source-id: 939a08c8e698358fdb75fb971ed509546ecbd7ea