generate chrome settings from a reference registry file
Summary:
This is another way to address the recent github issue re: the Windows chrome provider. Instead of manually maintaining a list of known settings we use the reference files provided by Google to generate a module that has a list of known settings.

A new attribute is added to control gating, since our default recipe calls the `cpe_chrome` resource and there isn't an easy way to override the behavior when adding a new resource.

This new behavior will *actually do the right thing*. For instance, Joe added support for the `RelaunchNotification` setting. However, the old code makes a lot of assumptions about where this key should be placed and writes to a location in the registry that isn't picked up by the chrome executable. Using the approach with generated code we can always know *exactly* where to place settings since that information is supplied by the registry file.

Maintenance is also much simpler since we can grab the administrative templates from Google, update `win_chrome_policy.reg` and then run these commands from a devserver
```
$ ./scripts/windows_policy_gen.rb && arc lint --apply-patches
```

NOTE: The linter is drunk:
```
 ls -lah scripts/windows_policy_gen.rb
-rwxr-xr-x. 1 dansedlacek users 2.2K Mar 10 20:25 scripts/windows_policy_gen.rb
```

WARNING: windows_policy_gen.rb has the string `generated` so phabricator thinks it is also generated. You'll probably want to inspect it when you review this.

Reviewed By: nmcspadden

Differential Revision: D20380317

fbshipit-source-id: f9cf03e56ef013fa532a415896a8f70f8cc57967