filesystem2 => filesystem conversion: fb_helpers
Summary:
Converting the `fb_helpers` cookbook over to the `filesystem_data` node method.

In chef 13, the `filesystem` ohai plugin was replaced with the `filesystem2` plugin, and the `filesystem2` data is written to both the `filesystem2` and `filesystem` apis. All uses of the old `filesystem` api were converted to the `filesystem2` api before I upgraded the fleet to chef 13 (tracked in T28228267). In order to upgrade to Chef 14+, we need to remove the references to the `filesystem2` api. However, there are some outlier operating systems that missed the conversion and so we're going to use a node method to detect what api is available and use that one so we can continue to support this transition for those other platforms.

I'm also adding in some checks to validate that the various keys we are using in some of these methods actually exist so we don't end up with the `#<NoMethodError: undefined method `[]' for nil:NilClass>` issues

official deprecation: https://docs.chef.io/deprecations_ohai_filesystem.html
PR to covert BSD: https://github.com/chef/ohai/pull/1181
PR to convert the rest of Unix (solaris, aix): https://github.com/chef/ohai/pull/1266
PR to convert Windows: https://github.com/chef/ohai/pull/1267

Reviewed By: joshuamiller01

Differential Revision: D21030821

fbshipit-source-id: 5457a9cdc35c55a33f43413ad7f5cf20b8ea593d