Adding improved idempotency check
Summary:

Chrome does not behave well if the attributes are empty - it still creates an empty profile. This changes it so that the default attributes (which contain some empty lists and hashes as well as nil values) will not actually result in any change on the system, thus making this a more true API cookbook.

Reviewed By: svmastersamurai

Differential Revision: D21982902

fbshipit-source-id: 3ea81fdee123de7694745b2d9168ecb90dd4e1ab