Adding improved idempotency check
Summary:

cpe_spotlight is still takes actions even when using the default attributes, when there are no exclusions to add. There's no reason to create /var/chef in this cookbook, so that's being removed. This is now a more true API cookbook that is a total no-op when using the default attributes.

Reviewed By: svmastersamurai

Differential Revision: D21979135

fbshipit-source-id: 93373a923a4283dfec4e87d63991919c15e4d9d1