Better Windows safety
Summary:

This adds in a retry as well as ignore failures for the Windows hosts file. If it's in use at all when Chef tries to touch it, that causes Chef to explode because it won't have permission to overwrite it. Adding a retry lets us get a pretty good chance of actually doing it; failing that, we simply ignore it and move on without breaking the run.

This also removes the user/group/mode from the Windows portion, let's let Windows ACLs handle this the correct way by default.

Reviewed By: svmastersamurai

Differential Revision: D22288162

fbshipit-source-id: 8f57d27b74d76ff6bacbfe7eaeaa8e4e5d9c8545