New cookbook (replacement)
Summary:
New (replacement) cookbook: `cpe_profiles`. The previous `cpe_profiles` has been renamed to `cpe_profiles_local` (in previous diff in stack).

The new `cpe_profiles` implements a sort of "meta" profiles API (very similar/the same) as before. See the `README.md` for an overview of this it works.

But the tl;dr is, essentially, this `cpe_profiles` reads in profiles set in its API and re-writes them out for another cookbook to do the *actual* profile installation. In its default config it works with `cpe_profiles_local` which installs profiles using the `osx_profiles` Chef resource.

Reviewed By: chilcote

Differential Revision: D22386869

fbshipit-source-id: 615101b69a4f8015f84432b03fcb32a732ee05fa