Performance improvement for rule update
Summary:

We have changed the mechanism to generate uuids for the applocker rules in a way that we produce the same uuid when no meaningful rule changes have occurred. This has the effect that no rule update is performed client side when there is no need for it. For more background, see "Misc notes for adding new rules" in https://fburl.com/wiki/2wy6udq3. Specifically, the comment about preserving the order in which we push the rules to applocker from the Chef infrastructure.

Reviewed By: svmastersamurai, ameyah

Differential Revision: D22425795

fbshipit-source-id: 7bd2deb3e57da769d29477194bd4c67d2dc00e56