ohai hardware output
Summary:

it appears that apple has changed the output of system_profiler to no longer include remaining charge.

Since we don't use that value anywhere, we can just remove it. (Also, it's worth pointing out that the value in our ohai was meaningless anyway, and returns 0 for my mac)

However, this will also need to be fixed upstream. The only call site for the battery info is in the powermanagement resource, and we can use stock ohai data for that.

Reviewed By: svmastersamurai

Differential Revision: D22962356

fbshipit-source-id: 0f970acaff6b4652f6a6d7c032ba1d44cb2136ba