Update in_timeshard? to support a zero length duration
Summary:

Currenlty the minimum duration that can be specified with in_timeshard?
is an hour. So any change being rolled out will start at the specified
time and slowrly be rolled out via sharing over an hour. We'd like to be
able to roll out changes at a specified time without any additional
sharding, so let's add support for a zero length duration.

Reviewed By: nikitakoshikov

Differential Revision: D24069788

fbshipit-source-id: 74b3a44404fb7103b31091b8afb13bdb9b11d64f