allow ignoring failures on individual packages
Summary:

`cpe_flatpak`'s default is to not ignore failures (so failing to install a package would cause Chef to fail). We want to keep this but allow individual packages to fail (so if a feature like zoom uses `cpe_flatpak` to install software, we can make it not fail Chef).

Flathub is normally quite stable (but people can try adding more remotes with different failure properties) but there are cases, likely when a package is being updated, when failure is possible.

Reviewed By: steelcowboy

Differential Revision: D25625231

fbshipit-source-id: cbfc18a73594d212b83c975154f9d6ff8b6efa7f