make `rpm_cmpver` handle Arch Linux's sudo 1.9.5.p2-1
Summary:

We coax `Gem::Version` into comparing post-release lettered versions properly by replacing post-release
letters with a dot followed by a numerical value of that post-release letter (e.g. 1.9.5p2 -> 1.9.5.15.2).

This breaks on Arch where that version is prefixed with a dot; 1.9.5..15.2 is not a valid version.

Handle this by replacing double-dots in the result with a single dot.

Reviewed By: compiler-errors

Differential Revision: D26414146

fbshipit-source-id: 60d44bd62cdace149e314ebc604d43821da3c00e