fb_helpers: add a new FB::Helpers.group_exist? method
Summary: Similar method to `FB::Helpers.user_exist?` this method will check if a group has been defined on the system via the native ruby library. This will help users who need to gate behavior on if a user exists in the system but don't want to manually reload ohai to ensure they have up to date information.

Reviewed By: joshuamiller01

Differential Revision: D29153595

fbshipit-source-id: 4d2b7aa5d5bf1d0c2fca0d84458f1513a756b2b9