Change hardcoded 'root / root' in cookbooks that macOS uses to node object.
Summary:
This diff will change all hardcoded root/root ownerships and use the appropriate node object methods instead.
This way the cookbooks can be used on macOS (or possibly Windows!) as well without trying to forcefully create a group called 'root' and giving it UID 0. This is what we're currently doing but have to get away from. On macOS, uid 0 is reserved for the 'wheel' group.

Reviewed By: malmond77

Differential Revision: D29234520

fbshipit-source-id: 620eadef11463ff7d9414bf1dbb10fab53e2d14b