update node.rpm_version for chef 14+
Summary: After transitioning to `yum_helper` in Chef 14+ the output of the yum cache method `installed_version` has changed from `version` to `epoch.version.architecture`. Since we're already using the python helper with the dnf cache query, I'm updating this method to use the yum python helper for Chef 14+ and to parse the output (which is `epoch.version` for both dnf and yum) correctly. This correctly handles any `noarch` packages since the architecture has already been parsed for us, as opposed to the yum cache which requires us to detect and parse it, which can be challenging considering the many variations on package version conventions.

Reviewed By: joshuamiller01

Differential Revision: D29271443

fbshipit-source-id: c69159ae6774dd9366f026f43bd068e4c2fa7068