fix chrome intial preference location
Summary:

In June 2020, Chromium landed a diff which changed the install location for Chromium (and Google Chrome some time not too much later).  Historically, Chrome had used C:\Program Files (x86)\ even for x86_64 versions.  Now they continue to use that location for existing installs which upgrade to new versions, but use C:\Program Files\ for new installs.

In addition, Chrome is migrating from the master_preferences file name to initial_preferences, presumably to the biased terminology of master.

This diff replaces writing to one location with writing to all 4 locations, to ensure that we'll be writing the file for all places chrome might be looking for it.

Reviewed By: steelcowboy

Differential Revision: D30918324

fbshipit-source-id: 38fe87842fa6156eacb865b2aeb569ac278474cc