add support for product_version wufb setting
Summary:
It *could* be that our current settings aren't specific enough to prevent a Windows 11 upgrade. This option will allow recipes to specify a Windows release for version targeting.

To head off any FUD there seems to have been only one instance of this in the last 30 days: https://fburl.com/scuba/cpe_windows_update_client/gf5o3k8y
And the number of Windows 11 devices checking into Defender is 15, so this is not a widespread problem from the data available.

Reviewed By: rslowe

Differential Revision: D32073765

fbshipit-source-id: 7b0d25d98847ecc2c8c987afe8ef289322219060