Handle username lookup correctly on windows
Summary:
```
```

The way that username is looked up was changed in D23615399 (https://github.com/facebook/IT-CPE/commit/84e14783ca4ba02a66edc33bad7b7d71e1f80e88) but there are some case where a host has a different language set which caused an array to be return which is breaking chef.  For example
```
[" eagbor                console            28  Actif       aucun   07/02/2022 19:10\r"]
```

this causes chef to error plus the report handlers fails which prevent us from collecting those failed runs.  In about 4 hours,  there was ~50 chef errors

{P478918658}

To fix this issue,  I'm combining the new and the previous method to look up the username.   if the new method returns nil,  the old method is used to lookup the username from the registry.

Reviewed By: HeyItsGilbert

Differential Revision: D34060001

fbshipit-source-id: a344f5022767c2e20191e43ca55b8792362624f5