Setting WuFB's Target OS Version to 21H2 for CPE
Summary:

Data Security Engineering is eyeing a registry key for external storage protection that’s only available in Windows 10v2106 (21H2).  So we're testing the target OS for Windows systems for upgrades to 21H2 from the current 21H1 it's set to now.

This diff is to only going to CPE Windows users first to verify the update doesn't break anything.

Differential Revision: D34889982

fbshipit-source-id: 4fc0f3ae2aff3d303c9c1395956ec14387f8f379