Add helper method for creating lazy attributes
Summary:

Chef 17 added support for automatically calling DelayedEvaluators when they are read from the node object. This diff creates a helper method to safely start leveraging this to remove runtime dependencies on clients that are >= chef 17 without directly breaking older clients.

Using this helper method should be done with with caution since it will result in different behaviors on different client versions.

Reviewed By: joshuamiller01

Differential Revision: D36212417

fbshipit-source-id: 7cd5fbafa939a02afba49431f2ada75777272df0