add regkey to wufb resource to support toggling target releases on or off
Summary: According to the interwebs there is [an additional key](https://admx.help/?Category=Windows_10_2016&Policy=Microsoft.Policies.WindowsUpdate::TargetReleaseVersion) the resource is not tracking. This change will allow the consumer of the cookbook to toggle the feature on or off if the feature is supported on that operating system release.

Reviewed By: dtheyer

Differential Revision: D36294573

fbshipit-source-id: 4961b69519db6c4a22682c27e344e018bc4f66e3