Install bunnylol
Summary:
```
```

Chrome 102 is forcing users to log into a managed chrome profile which is not applying the bunnylol policy that we set for new profile creation.  https://fb.workplace.com/groups/gsuite.feedback/permalink/1069845367215983/

Given that we don't want to manage this user policy on all devices,  I'm adding a feature to allow people to "install" bunnylol without having to manually set the values.

Reviewed By: dtheyer

Differential Revision: D36953528

fbshipit-source-id: 79b94aa817e85d9bcd5db3bf4ca1b41eb791d8dc