Enable cpe_chrome to manage registry settings using a reg file on windows
Summary:

Chrome management consists of a significant number of registry keys and values that need to be set which can take a significant amount of time when using the `registry_key` resource. Using a reg file instead results in a dramatic speed improvement for clients that have a large number of policies configured.

Reviewed By: HeyItsGilbert

Differential Revision: D38092507

fbshipit-source-id: 01343df2c7d202c86612f31962946e30d58361ef