Allow idempotent nil checksums
Summary:
Currently if a nil checksum is passed in, it is blindly passed to `remote_file` which fails since nil is not a valid type for that field.

Now, when a nil checksum is passed, the cpe_remote_file idempotency check will only verify that the file exists, has the correct permissions, and the correct owners. The impact of this is that nil checksums are allowed, but files will never be overriden if the remote version changes.

Reviewed By: gbatye

Differential Revision: D40986953

fbshipit-source-id: 347232467f494eb91405d6507474fcd4983d6f5d