fb_helpers: make unit testing gated_template resource easier
Summary:

Unit tests against anything which uses the gated_template resource are difficult.
This changes the way the resource is handled slightly to ease testing.
With that fix, some of the previous hacks to make this spec possible are no longer necessary,
so clean up the UserID crap

Within the custom resource, this switches from `converge_by` to use a template resource.  This should
be equivalent (if not DRY), but is what makes the unit testing work properly.  I tested the usual flow
for this via taste-tester, to confirm:

- Within permission, the file is not changed, and permission is requested
- With permission, the file is changed, and permission request is cleaned up
- With no change required, nothing happens

Differential Revision: D46749176

fbshipit-source-id: 1f0bd74d2a2929dcf62841b94f9bfa182f660369