do not manage driver update search on devices that have intune policy that manages it
Summary:
This change will permit us to specify `nil` for driver updates. We are currently experimenting with WufB driver updates via Intune and need to specify a way for Chef to backoff of managing the setting.

Internally we override this setting in another recipe scoped to a very small subset of devices when the Intune policy conflicts with what Chef is doing
```

  if registry_key_exists?('HKEY_LOCAL_MACHINE\SOFTWARE\Microsoft\MDMWins\device\Software/Policies/Microsoft/Windows/WindowsUpdate\ExcludeWUDriversInQualityUpdate')
    node.default['cpe_windows_update_for_business']['exclude_wu_drivers_in_quality_update'] = nil
  end
```

Differential Revision: D47927849

fbshipit-source-id: 9e96cc0e60a012ea513ddd4b7a5de26d281b6a11