Update PreFlight Script to remove SSID Check
Summary:
IOReg can on occasion return no information about what Wireless Network a device may be connected to.
Meta's CPE Team has determined that this check is not as useful as it was, so we are removing this check entirely from our own scripts.

Removing invalid shard_over_a_week_starting since it's in the past.

Differential Revision: D61231175

fbshipit-source-id: 6e7adcedd2f769d44edacd3afae3c4a38a49e1f9