Run NoOwnerGroupRoot on the entire codebase: itchef/cpe
Summary:
Just run the previous diff on the `itchef` code base to clean everything up: split the cpe cookbooks

```
$ hg show | grep '^[-+] ' | grep -vEe "^- *(owner|group) *'root'\$" -e  '^\+ *(owner|group) *node.root_(user|group)$'
-      has_elantech ||= get_dmesg('psmouse').include?('elantech')
+      elantech ||= get_dmesg('psmouse').include?('elantech')
+        owner ::Chef.node.root_user
+        group ::Chef.node.root_group
-      owner 'root' # rubocop: disable Chef/Meta/NoOwnerGroupRoot
-      owner 'root' # rubocop: disable Chef/Meta/NoOwnerGroupRoot
```
So it literally just changes what it's supposed to. Those lines show up because they're in libraries so foodcritic said I had to do this... I guess I didn't have the `node` object handy there.

The `elandtech` change is because of another lint that was complaining as the method is called `elandtech?`

Differential Revision: D56219064

fbshipit-source-id: 8b57a632d706d8a0c4177dfd01659b3bed29c563