include_recipe_at_converge_time resource
Summary:
In the course of migrating some resources to JSON recipes, it became clear that
many resources *could* be moved to JSON if there was some way of delaying the
resources to converge time when an attribute can be read (a common feature in
the FB API model). This new resource delays inserting the recipe into the run
context until converge time.

Differential Revision: D82673994

fbshipit-source-id: 66a88860331b21eee022dc752ba3d91bb284626e