Allow SelectedUserSID to be used in case LastLoggedInUserSID is Missing
Summary:
CPE @ Meta has experienced a possible edge case on filling the SID in case the User has only logged in via Windows RDP (via `mstsc` or the Windows App on MacOS)

If a machine is provisioned using a "defaultuser0" account, then logged into via RDP, the LastLoggedOnUserSID field will not be available until the user logs in via another Method, or reconnects the RDP user session. In this case, the person attribute will be filled out, with the exception of the uid field (which will be null).

In this case, the SelectedUserSID (which represents the current user) should be used. Defer to the LastLoggedOnUserSID whenever that is filled in.

Differential Revision: D86467726

fbshipit-source-id: 6f1ccd5ad389f04bcf813f615f0dd156eb9acc9c